requirements.txt
# pip install -r requirements.txt
scipy==1.12.0
Flask==3.0.2
opencv-python-headless==4.9.0.80
pillow==10.2.0
dropbox==11.36.2
pandas==2.2.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
matplotlib==3.8.2
timm==0.9.12
psutil==5.9.8